requirements.txt
torch>=1.7.1
transformers>=4.5.0
tensorflow
tensorboard
tqdm
fugashi
unidic
datasets>=1.13.3
sentencepiece>=0.1.91
chardet
keras==2.6
beautifulsoup4

japanize-matplotlib
janome

gensim==3.8.3
numpy==1.18.5
wordcloud==1.8.0
spacy==2.3.2
spacy_streamlit==0.1.0
seaborn==0.11.0
joblib==0.17.0
sumy==0.8.1
pandas==1.1.3
matplotlib
Pillow
scikit_learn==0.24.1